CR1 - create app/config and functions re-order
